+++ burn-wgpu/src/kernel/prng/mod.rs
@@ -1,3 +1,4 @@
 mod base;
+mod uniform;
 
-pub use base::*;
+pub use uniform::*;
