+++ burn-wgpu/src/ops/float_ops.rs
@@ -1,5 +1,5 @@
 use super::{numeric, BoolTensor, Device, FloatElem, FloatTensor, FullPrecisionBackend, IntTensor};
-use crate::kernel::prng::random_default;
+use crate::kernel::prng::random_uniform;
 use crate::kernel::{
     self, unary_default, unary_inplace_default, unary_scalar_default, unary_scalar_inplace_default,
 };
@@ -34,18 +34,20 @@ where
         device: &Device<Self>,
     ) -> FloatTensor<Self, D> {
         // TODO other distributions than default
-        if let Distribution::Default = distribution {
-            random_default::<G, F, D>(shape, device)
-        } else {
-            let mut seed = SEED.lock().unwrap();
-            let mut rng = if let Some(rng_seeded) = seed.as_ref() {
-                rng_seeded.clone()
-            } else {
-                get_seeded_rng()
-            };
-            let tensor = Self::from_data(Data::random(shape, distribution, &mut rng), device);
-            *seed = Some(rng);
-            tensor
+        match distribution {
+            Distribution::Default => random_uniform::<G, F, D>(shape, device, 0.elem(), 1.elem()),
+            Distribution::Uniform(low, high) => random_uniform::<G, F, D>(shape, device, low, high),
+            _ => {
+                let mut seed = SEED.lock().unwrap();
+                let mut rng = if let Some(rng_seeded) = seed.as_ref() {
+                    rng_seeded.clone()
+                } else {
+                    get_seeded_rng()
+                };
+                let tensor = Self::from_data(Data::random(shape, distribution, &mut rng), device);
+                *seed = Some(rng);
+                tensor
+            }
         }
     }
